requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0"
contourpy==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
crispy-bootstrap5==2024.2 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
django-components==0.84 ; python_version >= "3.12" and python_version < "4.0"
django-crispy-forms==2.2 ; python_version >= "3.12" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.12" and python_version < "4"
django-extensions==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
django-htmx==1.18.0 ; python_version >= "3.12" and python_version < "4.0"
django-lifecycle==1.2.4 ; python_version >= "3.12" and python_version < "4.0"
django-taggit==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
django-template-partials==24.2 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.7 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.12" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
faker==26.0.0 ; python_version >= "3.12" and python_version < "4.0"
fonttools==4.53.1 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.9.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
wordcloud==1.9.3 ; python_version >= "3.12" and python_version < "4.0"
